market-trends-agent/requirements.txt
strands-agents
boto3
playwright
bedrock-agentcore